🚨 fix lint: rename scripts/lib files in camelCase
