+++ scripts/lib/browserSdkVersion.js
