+++ scripts/lib/filesUtils.js
