+++ scripts/lib/packagesDirectoryNames.js
