+++ scripts/test/export-test-result.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { printLog, runMain } = require('../lib/execution-utils')
+const { printLog, runMain } = require('../lib/executionUtils')
 const { command } = require('../lib/command')
 const { getOrg2ApiKey } = require('../lib/secrets')
